Fix issue truncating spice coverage start/end date to day boundary (#1515)
* Fix issue truncating spice coverage start/end date to day boundary
Improve accuracy of conversion to J2000 seconds

* Add docstring to _seconds_since_j2000 function

* Apply suggestions from code review


* Allow for floating point times in input to metakernel api

* Copilot feedback

* Back out kernel furnishing change

---------

+++ sds_data_manager/lambda_code/SDSCode/api_lambdas/spice_query_api.py
@@ -66,9 +66,13 @@ def lambda_handler(event, context):
                 return response
             try:
                 if param == "start_time":
-                    query = query.where(models.SPICEFiles.max_date_j2000 >= int(value))
+                    query = query.where(
+                        models.SPICEFiles.max_date_j2000 >= float(value)
+                    )
                 elif param == "end_time":
-                    query = query.where(models.SPICEFiles.min_date_j2000 <= int(value))
+                    query = query.where(
+                        models.SPICEFiles.min_date_j2000 <= float(value)
+                    )
                 elif param == "type":
                     query = query.where(models.SPICEFiles.kernel_type == value)
                 elif param == "file_name":
